Add initial API endpoints for consumers
The consumers should have a relatively sane API to call without
needing to deal with the complexity of some of the NWS API endpoint
data expecations.

Codifying this first so that we have a view of what consumers will
need to provide will probably lead to better results as we're not
working strictly as a passthrough from NWS API to consumer input.