Separate Consumer-facing API from network I/O
NwsDataLoader is the network "layer" and interacts with the NWS API over
the network. This is separate from the NwsLoader (should be renamed), in
that the NwsLoader is for use by the consumer.

The decision was made this way so that testing can be introduced for
parsing messages without making enormous amounts of requests to the NWS
API every time we decide to test new functionality.

That is, there will be a "unit test" project with sample response APIs
that will be loaded from filesystem and parsed, rather than requested
from the NWS on every test.