Finish parsing for NwsPoint from response
The "unit" test is not one, but it helped uncover missing pieces of the
point information in the response that wasn't being parsed for yet.

The goal of these types of tests is to
a. Figure out if any members are missing from the classes representing
   the responses
b. Serve as a quick behavior check for future class modifications